Introduce workaround to show something resembling a back button in iOS
See discussion in https://github.com/woocommerce/WooCommerce-Shared/issues/61

The problem seems to be that the JS bundle generated by `make
bundle-ios` does not include, or knows how to refer to, the assets that
the react-navigation library uses to render the back button in iOS (not
sure why they don't access the system provided button, but there must be
a reason).

I tried to change the wiring so that the assets became available, but
got nowhere.

This workaround shows a chevron-like character in the navigation label
to make up for the lack of button image.